workspace/discord_surface/requirements.txt
discord.py>=2.4,<3
requests>=2.31,<3

